+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Waiting for Release
 - Update "Usage this month" value on Open plugin settings or press Apply
 - Add: Compress images in Background
 - Fix: NPE when choose bad image
+- Add: Transparency chessboard under image